+++ CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pieces won't rebuild themselves infinitely when a different piece is selected during playback
 - Pieces won't build twice when a different piece is selected during playback
 
+### Changed
+
+- Samples are now hosted on a separate site (currently at samples.generative.fm)
+
+### Added
+
+- New piece: "Observable Streams"
+
 ## [0.2.1] - 2018-02-08
 
 ### Fixed
@@ -46,7 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Sound files will no longer be fetched and cached during service worker installation. They'll be cached once they are fetched for the first time. This significantly reduces cache usage since only one audio format is used per client.
 
-[0.2.1]: https://github.com/generative-music/site/compare/v0.2.0...v0.2.1
+[0.3.0][0.2.1]: https://github.com/generative-music/site/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/generative-music/site/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/generative-music/site/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/generative-music/site/compare/v0.1.0...v0.1.1